Rewrite section on file filtering
Experimentation shows that escaping is not necessary except when
escaping a character that has special meaning for the regular
expression engine. For example, the following should be used to match
the .git folder and its contents:

--cx-flow.zip-exclude=\.git/.*

Experimentation also shows that enclosing the regular expression in
quotes leads to these being considered part of the regular
expression(s).